# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Homeroom API
  version: 0.1.0
  description: |
    Contract shared by the Next.js web app and the Flutter app. Every route is scoped to the
    caller's school and role by Postgres row-level security (see backend/migrations); the
    "Roles" line on each operation says who is meant to call it, and the database enforces it.

    **Conventions**
    - Auth: `Authorization: Bearer <jwt>`. The API maps the token to `people.id` and sets `app.person_id`.
    - IDs are UUIDs. Dates are `YYYY-MM-DD`; timestamps are RFC 3339 UTC. Money is a decimal string in GHS minor-unit-safe form (`"1850.00"`), never a float.
    - Lists are cursor-paginated: `?limit=50&cursor=...` returns `{ items, next_cursor }`.
    - Errors use `application/problem+json`. A row the caller may not see is reported as **404**, not 403, so existence is not leaked. 403 is for a visible resource the caller may not change.
    - Writes from the offline apps send an `Idempotency-Key` header; replays return the original result.
servers:
  - url: https://api.homeroom.example/v1
  - url: http://localhost:3000/api/v1
security:
  - bearer: []
...
paths:
...
  /admin/overview:
    get:
      operationId: getAdminOverview
      tags: [ Overview ]
      summary: Admin home
      description: 'Roles: admin. Today''s attendance rate, unmarked registers, fees outstanding, recent announcements.'
      responses:
        '4XX': { $ref: '#/components/responses/Problem' }
        '5XX': { $ref: '#/components/responses/Problem' }
        '200':
          description: Overview
          content: { application/json: { schema: { $ref: '#/components/schemas/AdminOverview' } } }
  /teacher/today:
    get:
      operationId: getTeacherToday
      tags: [ Overview ]
      summary: 'Teacher home, periods for a day with register status'
      description: 'Roles: teacher.'
      parameters: [ { name: date, in: query, schema: { type: string, format: date }, description: Defaults to today in the school's timezone } ]
      responses:
        '4XX': { $ref: '#/components/responses/Problem' }
        '5XX': { $ref: '#/components/responses/Problem' }
        '200':
          description: Periods in order
          content:
            application/json:
              schema:
                type: object
                properties:
                  date: { type: string, format: date }
                  periods:
                    type: array
                    items: { $ref: '#/components/schemas/TodayPeriod' }
                required: [ date, periods ]

  # ------------------------------------------------------------------ students
...
  /class_sections/{id}/attendance:
    parameters: [ { $ref: '#/components/parameters/Id' } ]
    get:
      operationId: getRegister
      tags: [ Attendance ]
      summary: The register for one period and date
      description: 'Roles: teacher (taught), admin. Returns every enrolled student, with `status: null` where nothing is marked yet.'
      parameters:
        - { name: date, in: query, required: true, schema: { type: string, format: date } }
        - { name: period_id, in: query, schema: { type: string, format: uuid } }
      responses:
        '4XX': { $ref: '#/components/responses/Problem' }
        '5XX': { $ref: '#/components/responses/Problem' }
        '200': { description: 'Register', content: { application/json: { schema: { $ref: '#/components/schemas/Register' } } } }
    put:
      operationId: submitRegister
      tags: [ Attendance ]
      summary: Mark a whole period at once (also the offline-sync endpoint)
      description: |
        Roles: teacher (taught, term open), admin. Upserts on (student, section, period, date), so it is safe to replay.
        Each entry is validated independently and the response reports every outcome, so one bad row
        (for example a student no longer enrolled) does not lose the rest of a queued offline register.
        Conflict rule: **last write wins by `marked_at`**; an older queued write never overwrites a newer one.
        Triggers `attendance.marked` for each absent or late student (push, else SMS).
      parameters: [ { $ref: '#/components/parameters/IdempotencyKey' } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/RegisterSubmission' } } }
      responses:
        '4XX': { $ref: '#/components/responses/Problem' }
        '5XX': { $ref: '#/components/responses/Problem' }
        '200': { description: 'Per-entry outcomes', content: { application/json: { schema: { $ref: '#/components/schemas/RegisterResult' } } } }
        '403': { $ref: '#/components/responses/Problem' }
  /attendance_records/{id}:
    parameters: [ { $ref: '#/components/parameters/Id' } ]
    patch:
      operationId: correctAttendanceRecord
      tags: [ Attendance ]
      summary: Correct a single record
      description: 'Roles: teacher (taught, term open), admin (any term).'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties: { status: { $ref: '#/components/schemas/AttendanceStatus' }, note: { type: string, nullable: true } }
      responses:
        '4XX': { $ref: '#/components/responses/Problem' }
        '5XX': { $ref: '#/components/responses/Problem' }
        '200': { description: 'Updated', content: { application/json: { schema: { $ref: '#/components/schemas/AttendanceRecord' } } } }
  /schools/{id}/attendance_report:
    parameters: [ { $ref: '#/components/parameters/Id' } ]
    get:
      operationId: getAttendanceReport
      tags: [ Admin ]
      summary: Date-ranged attendance register export
      description: 'Roles: admin. Per-pupil counts over a date range. For a spreadsheet use the .csv variant.'
      parameters:
        - { name: from, in: query, required: true, schema: { type: string, format: date } }
        - { name: to, in: query, required: true, schema: { type: string, format: date } }
        - { name: class_section_id, in: query, schema: { type: string, format: uuid } }
      responses:
        '4XX': { $ref: '#/components/responses/Problem' }
        '5XX': { $ref: '#/components/responses/Problem' }
        '200':
          description: Report
          content:
            application/json: { schema: { $ref: '#/components/schemas/AttendanceReport' } }
  /schools/{id}/attendance_report.csv:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      operationId: getAttendanceReportCsv
      tags:
        - Admin
      summary: Attendance register export as a CSV file
      description: "Roles: admin. The same report as a download: one row per pupil, formula-safe, with a Content-Disposition filename."
      parameters:
        - name: from
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: class_section_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: The file
          content:
            text/csv:
              schema:
                type: string
        4XX:
          $ref: "#/components/responses/Problem"
        5XX:
          $ref: "#/components/responses/Problem"

  # ------------------------------------------------------------------ gradebook
...
components:
  securitySchemes:
    bearer: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
    Id: { name: id, in: path, required: true, schema: { type: string, format: uuid } }
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema: { type: string, maxLength: 100 }
      description: Client-generated; replays within 24 h return the original response.
  responses:
...
    Problem:
      description: Error
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
...
    Problem:
      type: object
      required: [ status, title ]
      properties:
        status: { type: integer }
        title: { type: string }
        detail: { type: string }
        code: { type: string, example: term_closed }
        errors:
          type: array
          items: { type: object, properties: { field: { type: string }, message: { type: string } } }
        request_id:
          type: string
          description: Matches the X-Request-Id response header and the server logs
    Money:
      type: string
      pattern: '^[0-9]+[.][0-9]{2}$'
      example: '1850.00'
    AttendanceStatus: { type: string, enum: [ present, late, absent, excused ] }
...
    Role: { type: string, enum: [ admin, teacher, guardian, student ] }
...
    StudentRef:
      type: object
      required: [ id, full_name ]
      properties:
        id: { type: string, format: uuid }
        full_name: { type: string }
        class_section: { type: object, properties: { id: { type: string, format: uuid }, name: { type: string } }, required: [ id, name ] }
...
    Person:
      type: object
      required: [ id, full_name, role ]
      properties:
        id: { type: string, format: uuid }
        full_name: { type: string }
        role: { $ref: '#/components/schemas/Role' }
        contact: { type: object, properties: { phone: { type: string }, email: { type: string } } }
        assignments:
          type: array
          description: 'Teachers, in the admin directory: what they teach, in the current term'
          items:
            type: object
            properties:
              class_section: { type: object, properties: { id: { type: string, format: uuid }, name: { type: string } }, required: [ id, name ] }
              subject: { type: string }
            required: [ class_section, subject ]
...
    AdminOverview:
      type: object
      properties:
        date: { type: string, format: date }
        attendance_rate_today_pct: { type: number, nullable: true }
        registers_unmarked:
          type: array
          items:
            type: object
            properties:
              class_section: { $ref: '#/components/schemas/ClassSection' }
              period: { $ref: '#/components/schemas/Period' }
              teacher: { $ref: '#/components/schemas/Person' }
            required: [ class_section, period, teacher ]
        fees_outstanding: { $ref: '#/components/schemas/Money' }
        invoices_overdue: { type: integer }
        recent_announcements: { type: array, items: { $ref: '#/components/schemas/Announcement' } }
      required: [ date, attendance_rate_today_pct, registers_unmarked, fees_outstanding, invoices_overdue, recent_announcements ]
...
    ClassSection:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string, example: 6A }
        grade_level: { type: string }
        term_id: { type: string, format: uuid }
        homeroom_teacher: { $ref: '#/components/schemas/Person' }
        subject: { type: string, description: 'Present on /me and /class_sections for a teacher: the subject they teach here' }
      required: [ id, name, grade_level, term_id ]
...
    Period:
      type: object
      properties:
        id: { type: string, format: uuid }
        subject: { type: string }
        teacher: { $ref: '#/components/schemas/Person' }
        weekday: { type: integer, minimum: 1, maximum: 7, description: 'ISO, 1 = Monday' }
        starts_at: { type: string, example: '08:00' }
        ends_at: { type: string, example: '08:45' }
      required: [ id, subject, teacher, weekday, starts_at, ends_at ]
...
    TodayPeriod:
      type: object
      properties:
        period: { $ref: '#/components/schemas/Period' }
        class_section: { $ref: '#/components/schemas/ClassSection' }
        register: { type: string, enum: [ unmarked, partial, complete ] }
        marked: { type: integer }
        enrolled: { type: integer }
      required: [ period, class_section, register, marked, enrolled ]
    AttendanceRecord:
      type: object
      properties:
        id: { type: string, format: uuid }
        student_id: { type: string, format: uuid }
        class_section_id: { type: string, format: uuid }
        period_id: { type: string, format: uuid, nullable: true }
        period_date: { type: string, format: date }
        status: { $ref: '#/components/schemas/AttendanceStatus' }
        note: { type: string, nullable: true }
        marked_by: { type: string, format: uuid }
        marked_at: { type: string, format: date-time }
      required: [ id, student_id, class_section_id, period_id, period_date, status, note, marked_by, marked_at ]
    Register:
      type: object
      properties:
        class_section_id: { type: string, format: uuid }
        period_id: { type: string, format: uuid, nullable: true }
        date: { type: string, format: date }
        term_closed: { type: boolean, description: 'True means read-only for teachers' }
        entries:
          type: array
          items:
            type: object
            properties:
              student: { $ref: '#/components/schemas/StudentRef' }
              status: { type: string, nullable: true, enum: [ present, late, absent, excused, null ] }
              note: { type: string, nullable: true }
              record_id: { type: string, format: uuid, nullable: true }
            required: [ student, status, note, record_id ]
      required: [ class_section_id, period_id, date, term_closed, entries ]
    RegisterSubmission:
      type: object
      required: [ date, entries ]
      properties:
        date: { type: string, format: date }
        period_id: { type: string, format: uuid, nullable: true }
        entries:
          type: array
          items:
            type: object
            required: [ student_id, status, marked_at ]
            properties:
              student_id: { type: string, format: uuid }
              status: { $ref: '#/components/schemas/AttendanceStatus' }
              note: { type: string }
              marked_at: { type: string, format: date-time, description: 'Device time of the tap; decides last-write-wins' }
    BatchResult:
      type: object
      properties:
        applied: { type: integer }
        rejected: { type: integer }
        results:
          type: array
          items:
            type: object
            properties:
              student_id: { type: string, format: uuid }
              outcome: { type: string, enum: [ applied, stale, rejected ] }
              code: { type: string, nullable: true, example: not_enrolled }
              message: { type: string, nullable: true }
            required: [ student_id, outcome ]
      required: [ applied, rejected, results ]
    RegisterResult:
      $ref: '#/components/schemas/BatchResult'
    AttendanceReport:
      type: object
      properties:
        from: { type: string, format: date }
        to: { type: string, format: date }
        rows:
          type: array
          items:
            type: object
            properties:
              student: { $ref: '#/components/schemas/StudentRef' }
              present: { type: integer }
              late: { type: integer }
              absent: { type: integer }
              excused: { type: integer }
              rate_pct: { type: number, nullable: true }
            required: [ student, present, late, absent, excused, rate_pct ]
      required: [ from, to, rows ]
...
    Announcement:
      type: object
      properties:
        id: { type: string, format: uuid }
        author: { $ref: '#/components/schemas/Person' }
        class_section_id: { type: string, format: uuid, nullable: true, description: 'Null = school-wide' }
        title: { type: string }
        body: { type: string }
        requires_response: { type: boolean }
        published_at: { type: string, format: date-time, nullable: true, description: 'Null = draft' }
        my_responses:
          type: array
          items: { $ref: '#/components/schemas/AnnouncementResponse' }
          description: 'Guardians, for slips that require a reply'
      required: [ id, class_section_id, title, body, requires_response, published_at ]
...
    AnnouncementResponse:
      type: object
      properties:
        announcement_id: { type: string, format: uuid }
        student_id: { type: string, format: uuid }
        guardian_id: { type: string, format: uuid }
        response: { type: string, enum: [ 'yes', 'no' ] }
        responded_at: { type: string, format: date-time }
      required: [ announcement_id, student_id, guardian_id, response, responded_at ]
...